Redefine overall code architecture
