+++ backend/src/observables/game/mod.rs
@@ -1,0 +1,3 @@
+pub mod controller;
+pub mod model;
+pub mod router;
