+++ backend/src/observables/grid/mod.rs
@@ -1,0 +1,1 @@
+pub mod model;
