+++ backend/src/observables/grid/model.rs
